# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cabeleleila Leila",
    "description": "Sistema de agendamento para o salão Cabeleleila Leila",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/config": {
      "get": {
        "tags": [
          "Configuração"
        ],
        "summary": "Get Config",
        "description": "Retorna a configuração atual (Super Admin apenas).",
        "operationId": "get_config_api_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemConfigResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Configuração"
        ],
        "summary": "Update Config",
        "description": "Atualiza a configuração do sistema (Super Admin apenas).",
        "operationId": "update_config_api_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SystemConfigResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "auto_approve_appointments": {
            "type": "boolean",
            "title": "Auto Approve Appointments"
          }
        },
        "type": "object",
        "required": [
          "id",
          "auto_approve_appointments"
        ],
        "title": "SystemConfigResponse"
      },
      "SystemConfigUpdate": {
        "properties": {
          "auto_approve_appointments": {
            "type": "boolean",
            "title": "Auto Approve Appointments"
          }
        },
        "type": "object",
        "required": [
          "auto_approve_appointments"
        ],
        "title": "SystemConfigUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}